Fix lint

+++ contracts/Counter.sol
@@ -2,13 +2,13 @@
 pragma solidity >=0.4.22 <0.9.0;
 
 contract Counter {
-  uint public count = 0;
+  uint256 public count = 0;
 
   function inc() public payable {
     count += count + 1;
   }
 
   function fail() public pure {
-    revert("Boom");
+    revert('Boom');
   }
 }
